+++ gameboy-core/src/main/kotlin/memory/io/sound/SoundOutput.kt
@@ -2,5 +2,5 @@ package memory.io.sound
 
 interface SoundOutput {
     fun reset()
-    fun play(left: Int, right: Int)
+    fun play(left: Byte, right: Byte)
 }
